Moves PageFooter component to components/layouts folder
