+++ layouts/default.vue
@@ -13,7 +13,7 @@
 import Vue from 'vue'
 import { Component } from 'vue-property-decorator'
 import QiskitOrgMenu from '~/components/layouts/QiskitOrgMenu.vue'
-import PageFooter from '~/components/footers/PageFooter.vue'
+import PageFooter from '~/components/layouts/PageFooter.vue'
 
 @Component({
   components: {
